* arts-2-1-545
* src/m_append.h: Several fixes.

Properly handle appending to or appending an empty variable.
Allow appending a variable to itself.
Don't run into an endless loop and eat all memory when appending an
ArrayOf to itself.

* src/mystring.h:

Fix bug in constructor that produced an assertion when trying to
create an arts String from an empty C++ string.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@8285 aaf1aab0-4228-0410-ad68-8dceda47f409