* arts-2-1-611
* src/lineshapes.cc: Made the lorentz function slower by calculating the
phase shape function as well, which is necessary for Zeeman and line mixing.

* src/absorption.cc (xsec_species_line_mixing_2nd_order): Fixed error in the
mathematics that ignored the first-order line-mixing. Also, it only accepts
calculations if the specified line-shape returns the phase, as this is 
necessary for line-mixing. Also, now takes data in SI units. The data
will soon be added to arts-xml-data.

-- This line, and those below, will be ignored--

M    ChangeLog
M    src/lineshapes.cc
M    src/absorption.cc


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@8392 aaf1aab0-4228-0410-ad68-8dceda47f409