* arts-2-1-808
* doc/uguide/absorption.tex:  Added a warning about using HITRAN04 
and HITRAN08 in conjunction with line mixing.  There appears to be
some quantum coding error in the older versions of HITRAN that causes
the line mixing data to be matched with the wrong lines.  Recommended
solution is to use HITRAN2012.  I am presently uncertain how/where to
warn the user in ARTS about this issue.



git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@8736 aaf1aab0-4228-0410-ad68-8dceda47f409