* arts-2-1-837
* controlfiles/CMakeLists.txt:  Added one slow line mixing test.

* controlfiles/artscomponents/linemixing/*:  There is one more test here
that shows how to set this up for the entire band.  I might add a test
later to show how to do Zeeman for one line but Line Mixing for the
entire band.  The other files are updated because I had forgotten that
artscat4 require all species to be defined.  This error was pointed out
to me by Jana.

* controlfiles/artscomponents/zeeman/*:  The files are updated because I
had forgotten that artscat4 require all species to be defined.
This error was pointed out to me by Jana.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@8797 aaf1aab0-4228-0410-ad68-8dceda47f409