* arts-2-1-881
* src/m_cloudbox.cc, src/methods.cc (ScatteringDoitMergeParticles1D):

New experimental method to speed up DOIT calculations.

This method pre-calculates a weighted sum of all particles per pressure
level before the actual DOIT calculation is taking place in
*ScatteringDoit*. It should be called directly after pnd_fieldSetup
(but after cloudbox_checkedCalc).

*pnd_field* is resized to [np, np, 1, 1]. Where np is the number
of pressure levels inside the cloudbox. The diagonal elements of
the new *pnd_field* are set to 1, all others to 0. Accordingly,
*scat_data_array* is resized to np. Each particle is the weighted sum
of all particles at this presssure level. This is an experimental
method currently only working for very specific cases. All particles
must be of the same type and all particles must share the same f_grid
and za_grid. And pha_mat_data, ext_mat_data and abs_vec_data must be
all the same size. This method can only be used with a 1D atmosphere.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@8862 aaf1aab0-4228-0410-ad68-8dceda47f409