* arts-2-2-15
* src/m_optproperties.cc (ScatteringDoitMergeParticles1D):

Removed old code. Fix from jana was correct. We don't have+want an
operator* overload for Numeric*Vector Multiplication. This problem was
caused by the one exception we made (operator* implementation for
scalar Vector*Vector multiplication). The Numeric is automatically
casted to a ConstVectorView and the following scalar product fails
because the two Vectors have different number of elements.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@8913 aaf1aab0-4228-0410-ad68-8dceda47f409