* arts-2-3-146
* controlfiles/artscomponents/doit/TestDOITprecalcInit.arts:

Revert the comparison precision back to 2e-1 after discussion with
Jana. The idea of this test is to show that even with a 0 initial guess
field the calculation still converges to the "same" result. Just with
more iterations. Therefore it should not fail when commenting out
setting the initial field.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@9194 aaf1aab0-4228-0410-ad68-8dceda47f409