* arts-2-3-151
* src/cloudbox.cc (find_cloudlimits):

Variable i was used uninitialized if cloudbox_margin was -1 leading to
wrong return values for upper cloudbox limit.

Adapted doxygen documentation for lower and upper parameters to
'param[in,out]'. Changed 'Determine' to 'Adjust' in function
description to make clear that the upper and lower parameters have to
be initialized already outside this function. Otherwise this function
will return wrong values.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@9200 aaf1aab0-4228-0410-ad68-8dceda47f409