* arts-2-3-171
* src/agenda_class.cc (Agenda::set_outputs_to_push_and_dup):

Only add output variables of agendas to the list of outputs for WSMs
that do execute a generic input agenda. Fixes crash because of an
uninitialized value. Added assertion to double-check that the generic
input agenda is not a controlfile generated agenda. Should never happen
because this is checked elsewhere.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@9226 aaf1aab0-4228-0410-ad68-8dceda47f409