* arts-2-3-196
* src/m_jacobian.cc (jacobianDoit):

Fix critical bug: In the loop where the dydx result vector is
calculated, the index was missing. Thus, instead of writing into the
right position in dydx, the whole vector was overwritten with a
constant value in every iteration. This resulted for, e.g., constant
results for all frequencies. Bug reported by Verena and Freddy.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@9269 aaf1aab0-4228-0410-ad68-8dceda47f409